djangogramm/requirements.txt
Django==4.2.8
django_taggit==5.0.1
easy_thumbnails==2.8.5
django-extensions==3.2.3
Faker==22.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
django-debug-toolbar==4.3.0
gunicorn==21.2.0
django-environ==0.11.2